students/k3340/Meshcheriakov_Daniil/Lr1/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg[binary]>=3.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
# passlib 1.7.4 ломается с bcrypt 4.1+ (ошибки при хешировании пароля)
bcrypt>=4.0.1,<4.1.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
email-validator>=2.0.0
